# [improve] Prometheus streaming parsing optimization enhancements