--- a/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/http/HttpCollectImpl.java
+++ b/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/http/HttpCollectImpl.java
@@ -627,32 +627,34 @@ public class HttpCollectImpl extends AbstractCollect {
     }
 
     private void parseResponseByPrometheusExporter(InputStream content, List<String> aliasFields, CollectRep.MetricsData.Builder builder) throws IOException {
-        Map<String, MetricFamily> metricFamilyMap = OnlineParser.parseMetrics(content);
+        String metrics = builder.getMetrics();
+        Map<String, MetricFamily> metricFamilyMap = OnlineParser.parseMetrics(content, metrics);
         if (metricFamilyMap == null || metricFamilyMap.isEmpty()) {
             return;
         }
-        String metrics = builder.getMetrics();
-        if (metricFamilyMap.containsKey(metrics)) {
-            MetricFamily metricFamily = metricFamilyMap.get(metrics);
-            for (MetricFamily.Metric metric : metricFamily.getMetricList()) {
-                Map<String, String> labelMap = metric.getLabels()
-                        .stream()
-                        .collect(Collectors.toMap(MetricFamily.Label::getName, MetricFamily.Label::getValue));
-                CollectRep.ValueRow.Builder valueRowBuilder = CollectRep.ValueRow.newBuilder();
-                for (String aliasField : aliasFields) {
-                    String columnValue = labelMap.get(aliasField);
-                    if (columnValue != null) {
-                        valueRowBuilder.addColumn(columnValue);
-                    } else if (CommonConstants.PROM_VALUE.equals(aliasField) || CommonConstants.PROM_METRIC_VALUE.equals(aliasField)) {
-                        valueRowBuilder.addColumn(String.valueOf(metric.getValue()));
-                    } else {
-                        valueRowBuilder.addColumn(CommonConstants.NULL_VALUE);
-                    }
+        MetricFamily metricFamily = metricFamilyMap.get(metrics);
+        if (null == metricFamily || CollectionUtils.isEmpty(metricFamily.getMetricList())) {
+            return;
+        }
+        for (MetricFamily.Metric metric : metricFamily.getMetricList()) {
+            Map<String, String> labelMap = metric.getLabels()
+                    .stream()
+                    .collect(Collectors.toMap(MetricFamily.Label::getName, MetricFamily.Label::getValue));
+            CollectRep.ValueRow.Builder valueRowBuilder = CollectRep.ValueRow.newBuilder();
+            for (String aliasField : aliasFields) {
+                String columnValue = labelMap.get(aliasField);
+                if (columnValue != null) {
+                    valueRowBuilder.addColumn(columnValue);
+                } else if (CommonConstants.PROM_VALUE.equals(aliasField) || CommonConstants.PROM_METRIC_VALUE.equals(aliasField)) {
+                    valueRowBuilder.addColumn(String.valueOf(metric.getValue()));
+                } else {
+                    valueRowBuilder.addColumn(CommonConstants.NULL_VALUE);
                 }
-                builder.addValueRow(valueRowBuilder.build());
             }
+            builder.addValueRow(valueRowBuilder.build());
         }
     }
+
 
     private void parseResponseByDefault(String resp, List<String> aliasFields, HttpProtocol http,
                                         CollectRep.MetricsData.Builder builder, Long responseTime) {