--- a/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java
+++ b/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java
@@ -81,6 +81,94 @@ public class OnlineParser {
         return metricFamilyMap;
     }
 
+    public static Map<String, MetricFamily> parseMetrics(InputStream inputStream, String metric) throws IOException {
+        Map<String, MetricFamily> metricFamilyMap = new ConcurrentHashMap<>(10);
+        try {
+            int i = getChar(inputStream);
+            while (i != -1) {
+                if (i == '#' || i == '\n') {
+                    skipToLineEnd(inputStream).maybeEol().maybeEof().noElse();
+                } else {
+                    StringBuilder stringBuilder = new StringBuilder();
+                    stringBuilder.append((char) i);
+
+                    // parse the metricName to filter.
+                    int next = parseMetricName(inputStream, stringBuilder).maybeSpace().maybeLeftBracket().noElse();
+                    String metricName = stringBuilder.toString();
+                    stringBuilder.delete(0, stringBuilder.length());
+
+                    // step2: Determine whether this metric should be parsed.
+                    if (metric.equalsIgnoreCase(metricName)) {
+                        parseMetricFromName(inputStream, stringBuilder, metricFamilyMap, metricName, next);
+                    } else {
+                        skipToLineEnd(inputStream).maybeEol().maybeEof().noElse();
+                    }
+                }
+                i = getChar(inputStream);
+                // To address the `\n\r` scenario, it is necessary to skip
+                if (i == '\r') {
+                    i = getChar(inputStream);
+                }
+            }
+        } catch (FormatException e) {
+            log.error("prometheus parser failed because of wrong input format. {}", e.getMessage());
+            return null;
+        }
+        return metricFamilyMap;
+    }
+
+    /**
+     * Start parsing the complete metric from the already parsed metric name.
+     */
+    private static CharChecker parseMetricFromName(InputStream inputStream, StringBuilder stringBuilder,
+                                                   Map<String, MetricFamily> metricFamilyMap,
+                                                   String metricName, int next) throws IOException, FormatException {
+        MetricFamily metricFamily;
+        MetricFamily.Metric metric = new MetricFamily.Metric();
+
+        if (!metricFamilyMap.containsKey(metricName)) {
+            metricFamily = new MetricFamily();
+            metricFamily.setMetricList(new ArrayList<>());
+            metricFamily.setName(metricName);
+            metricFamilyMap.put(metricName, metricFamily);
+        } else {
+            metricFamily = metricFamilyMap.get(metricName);
+        }
+        int i = next;
+        if (i == ' ') {
+            i = skipSpaces(inputStream).getInt();
+        }
+
+        List<MetricFamily.Label> labelList = new LinkedList<>();
+        metric.setLabels(labelList);
+        if (i == '{') {
+            parseLabels(inputStream, stringBuilder, labelList);
+            i = skipSpaces(inputStream).getInt();
+        }
+
+        stringBuilder.delete(0, stringBuilder.length());
+        stringBuilder.append((char) i);
+        i = parseOneDouble(inputStream, stringBuilder).maybeSpace().maybeEol().maybeEof().noElse();
+        metric.setValue(toDouble(stringBuilder.toString()));
+        if (i == '\n' || i == -1) {
+            metricFamily.getMetricList().add(metric);
+            return new CharChecker(i);
+        }
+
+        i = skipSpaces(inputStream).getInt();
+        stringBuilder.delete(0, stringBuilder.length());
+        stringBuilder.append((char) i);
+        i = skipOneLong(inputStream).maybeSpace().maybeEol().maybeEof().noElse();
+        if (i == '\n' || i == -1) {
+            metricFamily.getMetricList().add(metric);
+            return new CharChecker(i);
+        }
+        i = skipSpaces(inputStream).maybeEol().maybeEof().noElse();
+
+        metricFamily.getMetricList().add(metric);
+        return new CharChecker(i);
+    }
+
     private static class FormatException extends Exception {
 
         public FormatException() {