--- a/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java
+++ b/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java
@@ -108,7 +108,7 @@ public class OnlineParser {
                     stringBuilder.delete(0, stringBuilder.length());
 
                     // step2: Determine whether this metric should be parsed.
-                    if (metric.equalsIgnoreCase(metricName)) {
+                    if (metric.equals(metricName)) {
                         parseMetricFromName(inputStream, stringBuilder, metricFamilyMap, metricName, next);
                     } else {
                         skipToLineEnd(inputStream).maybeEol().maybeEof().noElse();