@tomsun28 2025-09-20
hi there has some ci error

```
Compilation failure: 
Error:  /home/runner/work/hertzbeat/hertzbeat/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java:[112,73] cannot find symbol
Error:    symbol:   variable metricFamilyMap
Error:    location: class org.apache.hertzbeat.collector.collect.prometheus.parser.OnlineParser
Error:  /home/runner/work/hertzbeat/hertzbeat/hertzbeat-collector/hertzbeat-collector-basic/src/main/java/org/apache/hertzbeat/collector/collect/prometheus/parser/OnlineParser.java:[127,16] cannot find symbol
Error:    symbol:   variable metricFamilyMap
Error:    location: class org.apache.hertzbeat.collector.collect.prometheus.parser.OnlineParser
Error:  -> [Help 1]
```